requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.0.4
django-extensions==3.2.3
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.4
pexpect~=4.8.0
prompt-toolkit==3.0.43
psycopg==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
wcwidth==0.2.13

PyJWT~=2.3.0
Faker~=24.8.0
pip~=22.0.2
attrs~=21.2.0
distro~=1.7.0
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
colorama~=0.4.4
setuptools~=59.6.0
PyYAML~=5.4.1
requests~=2.25.1
future~=0.18.2
PyGObject~=3.42.1
zipp~=1.0.0
olefile~=0.46
keyring~=23.5.0
python-dateutil~=2.8.1
bcrypt~=3.2.0